Implement palindrome function using recursion
Treat 1 length and 2 length strings as base cases. String of length 1
is always a palindrome whereas string of length 2 is a palindrome when
both characters are the same. In the recursive case, compare first and
last characters until they are not equivalent in which case string is
not a palindrome or until a base case is reached after shortening
string after each recursive call.